# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Stealth Console API
  version: 0.1.0
  description: Control-plane API for Stealth Console. Browser sessions are opaque HttpOnly cookies and are never returned in JSON.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/organizations/{organizationID}/projects:
    get:
      operationId: listProjects
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/OrganizationID' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses: { '200': { description: Projects scoped to an organization, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectsPage' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } }
    post:
      operationId: createProject
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/OrganizationID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateProjectRequest' } } } }
      responses: { '201': { description: Project created, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '409': { $ref: '#/components/responses/Conflict' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /v1/projects/{projectID}:
    get:
      operationId: getProject
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }]
      responses: { '200': { description: Project, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/NotFound' } }
    patch:
      operationId: updateProject
      description: Update mutable project metadata. Only project owners and admins may change the project name.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateProjectRequest' } } } }
      responses: { '200': { description: Updated project, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/ValidationError' }, '500': { $ref: '#/components/responses/InternalError' } }
    delete:
      operationId: deleteProject
      description: Permanently delete a project and all of its tenant-owned resources. Only the organization owner may perform this operation; confirm_name must exactly match the current project name.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DeleteProjectRequest' } } } }
      responses: { '204': { description: Project and all child resources deleted }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/ValidationError' }, '500': { $ref: '#/components/responses/InternalError' } }
...
  /v1/projects/{projectID}/sites/{siteID}/domains:
    get:
      operationId: listSiteDomains
      description: List custom hostnames bound to a Site. API-key callers require sites.read.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Cursor' }]
      responses: { '200': { description: Site domains, content: { application/json: { schema: { $ref: '#/components/schemas/SiteDomainsPage' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
    post:
      operationId: createSiteDomain
      description: Bind a DNS hostname to a Site and return the TXT challenge record. The hostname serves only after verification.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateSiteDomainRequest' } } } }
      responses: { '201': { description: Domain binding created, content: { application/json: { schema: { $ref: '#/components/schemas/SiteDomainResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' } }
  /v1/projects/{projectID}/sites/{siteID}/domains/{domainID}:
    get:
      operationId: getSiteDomain
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }, { $ref: '#/components/parameters/DomainID' }]
      responses: { '200': { description: Site domain, content: { application/json: { schema: { $ref: '#/components/schemas/SiteDomainResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      operationId: deleteSiteDomain
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }, { $ref: '#/components/parameters/DomainID' }]
      responses: { '204': { description: Domain binding deleted }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' } }
  /v1/projects/{projectID}/sites/{siteID}/domains/{domainID}/verify:
    post:
      operationId: verifySiteDomain
      description: Verify the required TXT record at _stealth-verification.hostname. Verified hosts may use the optional in-process ACME manager or an external TLS terminator.
      security: [{ sessionCookie: [] }, { projectApiKey: [] }]
      parameters: [{ $ref: '#/components/parameters/ProjectID' }, { $ref: '#/components/parameters/SiteID' }, { $ref: '#/components/parameters/DomainID' }]
      responses: { '200': { description: Verified domain, content: { application/json: { schema: { $ref: '#/components/schemas/SiteDomainResponse' } } } }, '400': { $ref: '#/components/responses/BadRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: stealth_session }
    projectAppCookie: { type: apiKey, in: cookie, name: stealth_app_<project-id-without-hyphens>, description: Project-specific opaque HttpOnly cookie. The actual name is derived from the path projectID; it is never the Console stealth_session cookie. }
    projectApiKey: { type: apiKey, in: header, name: X-Stealth-Key, description: Project-bound server-to-server secret. Supported scopes include users, databases, storage, functions, sites, webhooks, realtime.read, and messaging read/write pairs; never send this header through the Console browser bridge. }
  parameters:
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Cursor: { name: cursor, in: query, schema: { type: string, format: uuid } }
    OrganizationID: { name: organizationID, in: path, required: true, schema: { type: string, format: uuid } }
...
    ProjectID: { name: projectID, in: path, required: true, schema: { type: string, format: uuid } }
...
    SiteID: { name: siteID, in: path, required: true, schema: { type: string, format: uuid } }
...
    DomainID: { name: domainID, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
    BadRequest: { description: Invalid JSON or parameter, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    Forbidden: { description: Insufficient organization role, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: Resource conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    InternalError: { description: Internal server error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    PayloadTooLarge: { description: Request body exceeds the endpoint's configured limit (JSON requests are bounded separately from multipart uploads), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    UnsupportedMediaType: { description: Content-Type must be application/json except for file upload multipart/form-data endpoints, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    CreateProjectRequest: { type: object, additionalProperties: false, required: [name], properties: { name: { $ref: '#/components/schemas/Slug' } } }
    UpdateProjectRequest: { type: object, additionalProperties: false, required: [name], properties: { name: { $ref: '#/components/schemas/Slug' } } }
    DeleteProjectRequest: { type: object, additionalProperties: false, required: [confirm_name], properties: { confirm_name: { $ref: '#/components/schemas/Slug', description: Must exactly match the current project name. } } }
...
    Slug: { type: string, pattern: '^[a-z0-9][a-z0-9-]{1,62}$' }
...
    Project: { type: object, required: [id, organization_id, name, created_at], properties: { id: { type: string, format: uuid }, organization_id: { type: string, format: uuid }, name: { $ref: '#/components/schemas/Slug' }, created_at: { type: string, format: date-time } } }
...
    Pagination: { type: object, required: [limit, next_cursor], properties: { limit: { type: integer }, next_cursor: { type: [string, 'null'], format: uuid } } }
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string } } } } }
...
    ProjectResponse: { type: object, required: [project], properties: { project: { $ref: '#/components/schemas/Project' } } }
...
    ProjectsPage: { type: object, required: [projects, pagination], properties: { projects: { type: array, items: { $ref: '#/components/schemas/Project' } }, pagination: { $ref: '#/components/schemas/Pagination' } } }
...
    SiteDomain: { type: object, additionalProperties: false, required: [id, project_id, site_id, hostname, status, verification_token, verification_record_name, verification_record_type, verification_record_value, tls_status, created_at, updated_at], properties: { id: { type: string, format: uuid }, project_id: { type: string, format: uuid }, site_id: { type: string, format: uuid }, hostname: { type: string, pattern: '^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\.[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?)+$' }, status: { type: string, enum: [pending, verified, disabled] }, verification_token: { type: string, pattern: '^[0-9a-f]{64}$' }, verification_record_name: { type: string }, verification_record_type: { type: string, enum: [TXT] }, verification_record_value: { type: string }, verified_at: { type: [string, 'null'], format: date-time }, tls_status: { type: string, enum: [external, pending, active, failed] }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } } }
    CreateSiteDomainRequest: { type: object, additionalProperties: false, required: [hostname], properties: { hostname: { type: string, minLength: 4, maxLength: 253, description: Lowercase DNS hostname; IP addresses and wildcard names are not supported. } } }
...
    SiteDomainResponse: { type: object, additionalProperties: false, required: [domain], properties: { domain: { $ref: '#/components/schemas/SiteDomain' } } }
    SiteDomainsPage: { type: object, additionalProperties: false, required: [domains, pagination, can_manage], properties: { domains: { type: array, items: { $ref: '#/components/schemas/SiteDomain' } }, pagination: { $ref: '#/components/schemas/Pagination' }, can_manage: { type: boolean } } }
...